pathViewer: Fix possible misuse of GJS garbage collector
From logs:

Attempting to call back into JSAPI during the sweeping phase of GC.
This is most likely caused by not destroying a Clutter actor or Gtk+
widget with ::destroy signals connected, but can also be caused by
using the destroy(), dispose(), or remove() vfuncs. Because it would
crash the application, it has been blocked and the JS callback not
invoked.

This is quite hard to reproduce, so I won't be really sure
this makes any  difference until I can test it more.