filesystemOther: Fix another quirk of removing of negated filesystem
Say you have !~/test among the original permissions. Usually, that
would only make sense if there's a broader permission like home:ro
among the original permissions as well.

So, these negated permissions are only useful as exceptions to the
broader permission.

If !~/test is removed, the end result is that the exception is
removed and that  the broader permission takes prescedence, so
there is not a clear value on explictly displaying the ~/.test
override.