widgets: Always update path entries
Persistent permissions can't be negated. Therefore when a persistent
override is added globally, these overrides shouldn't be removed from
specific apps

What happens here  is that, Flatseal disables the remove  button for
the first app that is loaded and later, when switching to the global
overrides view, since both have the same persistent value the button
remains disabled.

Closes #631