models: Index permissions models
Before this change, to determine the appropriate permission model
for a given group+key+option tuple, it would iterate over each of
the models each time.

With this change that's no longer needed. Plus, it's now possible
to  apply  different  treatments in  cases where the  model don't
parse the value in the same way as the oher models,  e.g., envars
overrides.

Issue #613